requirements.txt
Babel==2.3.4
contextlib2==0.5.4
defusedxml==0.4.1
Django==1.10.5
django-allauth==0.30.0
django-allauth-bootstrap==0.2
django-appconf==1.0.2
django-autocomplete-light==3.2.1
django-bootstrap-form==3.2.1
django-compressor==2.1
django-compressor-toolkit==0.6.0
django-crispy-forms==1.6.1
django-environ==0.4.1
django-formtools==2.0
django-npm==1.0.0
django-phonenumber-field==1.1.0
django-solo==1.1.2
gunicorn==19.6.0
oauthlib==2.0.1
phonenumberslite==8.0.0
psycopg2==2.6.2
python-dateutil==2.6.0
python3-openid==3.0.10
pytz==2016.10
raven==5.32.0
rcssmin==1.0.6
requests==2.12.4
requests-oauthlib==0.7.0
rjsmin==1.0.12
six==1.10.0
whitenoise==3.2.3
